Split of benchmark and permutation engines
