+++ .gitignore
@@ -1,6 +1,6 @@
 # Generated by Cargo
 # will have compiled files and executables
-/target/
+target/
 
 # Remove Cargo.lock from gitignore if creating an executable, leave it for libraries
 # More information here https://doc.rust-lang.org/cargo/guide/cargo-toml-vs-cargo-lock.html
@@ -21,3 +21,7 @@ Cargo.lock
 *.pdb
 *.log
 /results.txt
+/old/
+
+# ignore the on-the-fly font file embedded in the executable
+tmp.flf
