+++ cli/src/lib.rs
@@ -1,0 +1,2 @@
+pub mod cli_util;
+pub mod supported;
