+++ engine/Cargo.toml
@@ -1,0 +1,16 @@
+[package]
+name = "engine"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+compound_duration = "1.2.0"
+itertools = "0.10.5"
+crossbeam-channel = "0.5.6"
+ctrlc = "3.2.4"
+indicatif = "0.17.2"
+stoppable_thread = "0.2.1"
+permutation = { path = "../permutation" }
+ffmpeg = { path = "../ffmpeg" }
