+++ engine/src/nvenc.rs
@@ -1,0 +1,7 @@
+pub fn get_nvenc_presets() -> Vec<&'static str> {
+    return vec!["p1", "p2", "p3", "p4", "p5", "p6", "p7"];
+}
+
+pub fn get_nvenc_tunes() -> Vec<&'static str> {
+    return vec!["hq", "ll", "ull"];
+}
