+++ permutation/Cargo.toml
@@ -1,0 +1,7 @@
+[package]
+name = "permutation"
+version = "0.1.0"
+edition = "2021"
+
+[dependencies]
+ffmpeg = { path = "../ffmpeg" }
