+++ permutor/Cargo.toml
@@ -1,0 +1,14 @@
+[package]
+name = "permutor"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+clap = { version = "4.0.32", features = ["derive"] }
+environment = { path = "../environment" }
+engine = { path = "../engine" }
+cli = { path = "../cli" }
+permutation = { path = "../permutation" }
+ffmpeg = { path = "../ffmpeg" }
