--- src/ffmpeg.rs
@@ -1,4 +1,0 @@
-pub mod args;
-pub mod report_files;
-pub mod ffprobe;
-pub mod metadata;
