--- src/permutations.rs
@@ -1,6 +1,0 @@
-pub mod permute;
-pub mod nvenc;
-pub mod result;
-pub mod fps_stats;
-pub mod h264_hevc_nvenc;
-mod resolutions;
